src/test/kotlin/org/carth/aoc21/day15/Day15Test.kt
package org.carth.aoc21.day15

import org.carth.aoc21.common.Resources.inputAsListOfString
import org.carth.aoc21.common.Resources.testInputAsListOfString
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.MethodOrderer
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Order
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestMethodOrder

private const val DAY = "day15"

class Day15Tests {

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day15PartOneTest {

        @Test
        @Order(1)
        fun `solve part one of sample`() {
            // Act
            val answer = Day15(testInputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(40, answer)
        }

        @Test
        @Order(2)
        fun `solve part one`() {
            // Act
            val answer = Day15(inputAsListOfString(DAY)).solvePartOne()

            // Assert
            assertEquals(398, answer)
        }
    }

    @Nested
    @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
    inner class Day15PartTwoTest {

        @Test
        @Order(1)
        fun `solve part two of sample`() {
            // Act
            val answer = Day15(testInputAsListOfString(DAY)).solvePartTwo()

            // Assert
            assertEquals(315, answer)
        }

        @Test
        @Order(2)
        fun `solve part two`() {
            // Act
            val answer = Day15(inputAsListOfString(DAY)).solvePartTwo()
            // Assert
            assertEquals(2817, answer)
        }
    }

}